requirements.txt
Django>=1.8
boto>=2.35.0
django-storages>=1.5.1
